修改了状态服务器。
git-svn-id: https://172.16.100.245:443/svn/yihunze@20 6ec6bfb8-d06b-6c47-a642-727affc0819e

+++ server/yihunzeServer/Mainlib/source/databaseInstance.cpp
@@ -99,7 +99,7 @@ u_long CppMySQLQuery::numRow()
 {
 	if(_mysql_res==NULL)
 		return 0;
-	return _mysql_res->row_count;
+	return (u_long)_mysql_res->row_count;
 }
 
 //--------------------------------------------------------------------------------
